backend/requirements.txt
fastapi
uvicorn[standard]
langgraph
pydantic
httpx
python-dotenv
websockets
langchain-core
langchain-openai
